Move servo serial creation to SerialFactory
